% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONT_BS_comparison_plots.R
\name{add_absolute_change_columns}
\alias{add_absolute_change_columns}
\title{Add Absolute Change to BEDMethyl Data}
\usage{
add_absolute_change_columns(comparison_bedmethyl)
}
\arguments{
\item{comparison_bedmethyl}{A data.table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth (integer)
\item percent_methylation: percentage of reads observed to be methylated
(numeric)
}}
}
\value{
A modified version of the \code{comparison_bedmethyl} data.table with
additional columns:
\itemize{
\item absolute_change_percent_methylation: absolute change in percent
methylation (numeric)
\item absolute_change_read_depth: absolute change in read depth (integer)
}

The returned data.table retains the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item ONT_read_depth: read depth for ONT (integer)
\item ONT_percent_methylation: percentage of methylated reads in ONT
(numeric)
\item BS_read_depth: read depth for BS (integer)
\item BS_percent_methylation: percentage of methylated reads in BS (numeric)
}
}
\description{
Adds two columns:
\itemize{
\item "absolute change in read depth"
\item "absolute change in percent methylation"
}

to a \code{comparison_bedmethyl} data table.
This is primarily for the purpose of using the
\code{create_read_depth_plot()} and \code{create_percent_comparison_plot()} functions.
}
