% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedmethyl.R
\name{bedmethyl_format}
\alias{bedmethyl_format}
\title{Format of bedmethl files}
\arguments{
\item{methylation_data}{A data table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}
}
\description{
The Expected format for bedmethyl files. This is just for
documentation purposes
}
