% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPG_robustness_plots.R
\name{calculate_nearby_methylation}
\alias{calculate_nearby_methylation}
\title{Calculate the Average Methylation for Nearby CpGs}
\usage{
calculate_nearby_methylation(
  methylation_data,
  chromosome,
  current_start,
  min_distance = 0,
  max_distance = 30
)
}
\arguments{
\item{methylation_data}{A data.table with columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}

\item{chromosome}{The chromosome that the CpG lies on (string)}

\item{current_start}{The start position of the CpG in question (integer)}

\item{min_distance}{The smallest distance a CpG can be away from another to
be called nearby. Integer valued, defaults to 0.}

\item{max_distance}{The largest distance a CpG can be away from another to
be called nearby. Integer valued, defaults to 30.}
}
\value{
The average percent methylation for 'nearby' CpGs to the input CpG
(numeric)
}
\description{
Given distances of what defines two CpGs being 'nearby' this
function calculates the average percent methylation for such CpGs.
}
\examples{
# Given the following input data:
      chr start   end   name read_depth strand percent_methylation
   <char> <num> <num> <char>      <num> <char>               <num>
1:   chr1    10    11      m         21      .                   4
2:   chr1    15    16      m        123      .                   33
3:   chr1    20    21      m         12      .                   21
4:   chr1    25    26      m         54      .                   50

calculate_nearby_methylation(methylation_data, "chr1", 10, 0, 5)
 33
calculate_nearby_methylation(methylation_data, "chr1", 20, 0, 10)
 25
calculate_nearby_methylation(methylation_data, "chr1", 20, 6, 10)
 27
}
