% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_5mc_5hmc.R
\name{combine_5mc_5hmc}
\alias{combine_5mc_5hmc}
\title{Combine the Signal from 5hmC and 5mC}
\usage{
combine_5mc_5hmc(comparison_bedmethyl)
}
\arguments{
\item{comparison_bedmethyl}{A data table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item ONT_read_depth: read depth in the ONT data(integer)
\item ONT_percent_methylation: percentage of reads observed to be methylated
in the ONT data(numeric)
\item BS_read_depth: read depth in the BS-Seq data(integer)
\item BS_percent_methylation: percentage of reads observed to be methylated
in the BS-Seq data (numeric)
}}
}
\value{
A comparitive bedmethyl file with merged 5mC and 5hmC signal
}
\description{
Merges methylation signal for both "h" and "m". This is useful
when comparing ONT data to WGBS data, as WGBS doesn't discern between 5mC
and 5hmC (unlike modifided base called ONT data).
}
\examples{
# Read in a comparative bedmethyl file
bedmethyl <- read_comparison_bedmethyl("path/to/bedmethyl.bed")

head(bedmethyl, 4)
      chr start   end mark_name ONT_read_depth ONT_percent_methylation
   <char> <num> <num>    <char>          <num>                   <num>
1:   chr1     0   200         h              6                       5
2:   chr1     0   200         m              6                       5
3:   chr1   200   400         h             99                      10
4:   chr1   200   400         m             99                      20
   BS_read_depth BS_percent_methylation
           <num>                  <num>
1:             0                      0
2:             0                      0
3:             0                      0
4:             0                      0

# Combine signal
bedmethyl <- combine_5mc_5hmc(bedmethyl)

head(bedmethyl, 2)
      chr start   end mark_name ONT_read_depth ONT_percent_methylation
   <char> <num> <num>    <char>          <num>                   <num>
1:   chr1     0   200         m              6                      10
2:   chr1   200   400         m             99                      30
   BS_read_depth BS_percent_methylation
           <num>                  <num>
1:             0                      0
2:             0                      0
}
