% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedmethyl.R
\name{comparative_bedmethyl_format}
\alias{comparative_bedmethyl_format}
\title{Format of Comparison Bedmethl Files}
\arguments{
\item{comparison_bedmethyl}{A data table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item ONT_read_depth: read depth in the ONT data(integer)
\item ONT_percent_methylation: percentage of reads observed to be methylated
in the ONT data(numeric)
\item BS_read_depth: read depth in the BS-Seq data(integer)
\item BS_percent_methylation: percentage of reads observed to be methylated
in the BS-Seq data (numeric)
}}
}
\description{
The expected format for comparison bedmethyl files. This is
just for documentation purposes
}
