% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_bed_files.R
\name{create_bins}
\alias{create_bins}
\title{Create genomic windows for a chromosome}
\usage{
create_bins(chromosome, chromosome_sizes, bin_size)
}
\arguments{
\item{chromosome_sizes}{a data.table detailing the length of each chromosome
with columns "chromosome name" and "chromosome size"}

\item{bin_size}{The desired size of each region (integer)}

\item{chromsome}{The name of the chromsome (string)}
}
\value{
A data.table with columns "chromosome",
"starting base pair position" and "end base pair position"
}
\description{
Given a chromosome and its size this creates a data.table
that details regions of a fixed length spanning the chromosome
}
\examples{
 create_bins("22", chromosome_sizes, 200)

 Generates a data.table of the form:
 chr22	0	200
 chr22	200	400
 chr22	400	600
 ...
}
