% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_bed_files.R
\name{create_blank_bed_files}
\alias{create_blank_bed_files}
\title{A wrapper for \code{create_bins()}}
\usage{
create_blank_bed_files(chromosome_sizes, bin_size)
}
\arguments{
\item{chromosome_sizes}{a data.table detailing the length of each chromosome
with columns "chromosome name" and "chromosome size"}

\item{bin_size}{The desired size of each region (integer)}
}
\value{
A data.table with columns "chromosome",
"starting base pair position" and "end base pair position"
}
\description{
Using a desired bin size, creates data.tables in the form of
a bed file for chromosomes 1 to X. Each row is the next subsequent window
equal to the bin size.\code{create_bins()}
}
\examples{
 create_bins("22", chromosome_sizes, 200)

 Generates a data.table of the form:
 chr22	0	200
 chr22	200	400
 chr22	400	600
 ...

}
