% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONT_BS_comparison_plots.R
\name{create_correlation_plot}
\alias{create_correlation_plot}
\title{Generate a Density Heatmap Correlation Plot}
\usage{
create_correlation_plot(
  comparison_bedmethyl,
  mark = c("m", "h"),
  read_depth_filter = 30
)
}
\arguments{
\item{comparison_bedmethyl}{A data.table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth (integer)
\item percent_methylation: percentage of reads observed to be methylated
(numeric)
}}

\item{mark}{The name of the mark to inspect ("m" or "h")}

\item{read_depth_filter}{A filter to use on the read depth in your bedmethyl
file. This is in place to reduce random fluctuations (common in low count
data)}
}
\value{
A plot (ggplot) of type \code{geom_bin2d} showing how the percent
methylation in BS and ONT compare
}
\description{
Using comparitive methylation data from BS-Seq and ONT, a
density heatmap is produced between the percent methylation seen at each
CpG site for both data types.
}
\details{
The most ideal plot would be one where all colouration is displayed on the
main diagonal (datasets agree with one another). This is assuming of course
that both datasets are from the same study/sample etc.
}
