% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPG_robustness_plots.R
\name{create_cpg_robustness_data}
\alias{create_cpg_robustness_data}
\title{Create Data for \code{create_cpg_robustness_plot()}}
\usage{
create_cpg_robustness_data(
  methylation_data,
  min_distance = 0,
  max_distance = 30
)
}
\arguments{
\item{methylation_data}{A data table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}

\item{min_distance}{The smallest distance a CpG can be away from another to
be called nearby. Integer valued, defaults to 0.}

\item{max_distance}{The largest distance a CpG can be away from another to
be called nearby. Integer valued, defaults to 30.}
}
\value{
A modified version of the input data.table with the added column:
\itemize{
\item surrounding_methylation: Average methylation signal of 'nearby' CpGs
(numeric)
}
}
\description{
Adds the column: average methylation of surrounding CpGs to the
inputted methylation data.
}
\examples{
create_cpg_robustness_data(bedmethyl_data, 0, 5)

      chr start   end   name read_depth strand percent_methylation
   <char> <num> <num> <char>      <num> <char>               <num>
1:   chr1    10    11      m         21      .                   2
2:   chr1    15    16      m        123      .                   3
3:   chr1    20    21      m         12      .                   4
4:   chr1    25    26      m         54      .                   5
   surrounding_methylation
                     <num>
1:                       4
2:                       5
3:                       2
4:                       3
}
