% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPG_robustness_plots.R
\name{create_cpg_robustness_plot}
\alias{create_cpg_robustness_plot}
\title{Create a Plot Showcasing Methylation Stability}
\usage{
create_cpg_robustness_plot(methylation_data, min_distance, max_distance)
}
\arguments{
\item{methylation_data}{A data.table with columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}

\item{min_distance}{The smallest distance a CpG can be away from another to
be called nearby. Integer valued.}

\item{max_distance}{The largest distance a CpG can be away from another to
be called nearby. Integer valued.}
}
\value{
A plot (ggplot) of type \code{geom_bin2d} showing how the methylation
signal of CpGs correlate with 'nearby' CpGs.
}
\description{
Generates a density plot that shows how similar the methylation
signal is for surrounding CpG sites for each CpG.
}
\details{
A plot where most points lie on the main diagonal is one where
nearby CpGs are generally very similar (in terms of methylation status)

It might seem odd that min_distance is a parameter here. This is mainly
here to give the user a little more control. If you want to see how
dissimilar far away CpGs are from each other, you would want to increase
this value to remove closer CpGs.
}
