% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate_plots.R
\name{create_error_rate_data}
\alias{create_error_rate_data}
\title{Generate Data for \code{create_error_rate_plot()}}
\usage{
create_error_rate_data(methylation_data, max_read_depth, percent_threshold)
}
\arguments{
\item{methylation_data}{A data.table with columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}

\item{max_read_depth}{The maximum read depth threshold to view (integer).
Some data sets have a single site with really high read depth, plotting
all the way up to this point is a waste of computation and won't give you
meaningful data (estimating an error rate using a single site). Defaults to
1000.}

\item{percent_threshold}{The maximum percent methylation to consider
(numeric). Lower values are recommended. Defaults to 5.}
}
\value{
A data table with columns
\itemize{
\item n -> The minimum read depth considered (integer)
\item error_rate -> The error rate estimated (numeric)
}
}
\description{
Create a data table with the data required for the
\code{create_error_rate_plot()} function.
}
\examples{
create_error_rate_data(methylation_data, 1000, 5)
        n error_rate
    <int>      <num>
 1:     1          0.6
 2:     2          0.54
 3:     3          0.52312
 4:     4          0.491
 5:     5          0.43221
 6:     6          0.4
 7:     7          0.3
 8:     8          0.123
 9:     9          0.1
10:    10          0.05
}
