% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate_plots.R
\name{create_error_rate_plot}
\alias{create_error_rate_plot}
\title{Generate a Plot of Error Rates}
\usage{
create_error_rate_plot(
  methylation_data,
  max_read_depth = 1000,
  percent_threshold = 5
)
}
\arguments{
\item{methylation_data}{A data table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}

\item{max_read_depth}{The maximum read depth threshold to view (integer).
Some data sets have a single site with really high read depth, plotting
all the way up to this point is a waste of computation and won't give you
meaningful data (estimating an error rate using a single site). Defaults to
1000.}
}
\value{
A plot (ggplot) of type \code{geom_point} showcasing how the estimated
error rate changes with the minimum read depth considered
}
\description{
Find how the predicted error rate for your methylation data
changes as the read depth and percent methylation thresholds change
}
\examples{
create_error_rate_plot(methylation_data, 1000, 5)

}
