% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONT_BS_comparison_plots.R
\name{create_percent_comparison_plot}
\alias{create_percent_comparison_plot}
\title{Generate a Histogram for the Change in Percent Methylation}
\usage{
create_percent_comparison_plot(
  comparison_bedmethyl,
  mark = "m",
  read_depth_filter = 30
)
}
\arguments{
\item{comparison_bedmethyl}{A data.table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth (integer)
\item percent_methylation: percentage of reads observed to be methylated
(numeric)
}}

\item{mark}{The name of the mark to inspect ("m" or "h")}

\item{read_depth_filter}{A filter to use on the read depth in your bedmethyl
file. This is in place to reduce random fluctuations (common in low count
data)}
}
\value{
A plot (ggplot) of type \code{geom_histogram} showing the distribution
of the absolute change in percent methylation between the datasets
}
\description{
Using comparative methylation data from BS-Seq and ONT, a
histogram is produced to showcase the distribution of the absolute change
in percentage methylation seen between each dataset.
}
\examples{
# Read in comparison_bedmethyl file
bedmethyl <- read_comparison_bedmethyl("path/to/bedmethyl.bed")

# Create histogram
create_read_depth_plot(bedmethyl)

}
