% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONT_BS_comparison_plots.R
\name{create_read_depth_plot}
\alias{create_read_depth_plot}
\title{Generate a Histogram for the Change in Read Depth}
\usage{
create_read_depth_plot(comparison_bedmethyl)
}
\arguments{
\item{comparison_bedmethyl}{A modified version of a comparison_bedmethyl (data.table) with columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item ONT_read_depth: read depth for ONT (integer)
\item ONT_percent_methylation: percentage of methylated reads in ONT
(numeric)
\item BS_read_depth: read depth for BS (integer)
\item BS_percent_methylation: percentage of methylated reads in BS (numeric)
\item absolute_change_percent_methylation: absolute change in percent
methylation (numeric)
\item absolute_change_read_depth: absolute change in read depth (integer)
}}
}
\value{
A plot (ggplot) of type \code{geom_histogram} showing the distribution
of the absolute change in percent methylation between the datasets
}
\description{
Using comparative methylation data from BS-Seq and ONT, a
histogram is produced to showcase the distribution of the absolute change
in read depth seen between each dataset.
}
\details{
Use \code{add_absolute_change_columns()} to obtain the necessary columns
in your comparison_bedmethyl data table for this function.
(see ?chrombinarize::add_absolute_change_columns)
}
\examples{
# Read in comparison_bedmethyl file
bedmethyl <- read_comparison_bedmethyl("path/to/bedmethyl.bed")

# Add absolute change columns
bedmethyl <- add_absolute_change_columns(bedmethyl)

# Create histogram
create_read_depth_plot(bedmethyl)

}
