% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_densely_methylated_bins.R
\name{determine_dense_bins}
\alias{determine_dense_bins}
\title{Determines Densely Methylated Regions}
\usage{
determine_dense_bins(bin_counts_file, beta_threshold = 0.001)
}
\arguments{
\item{bin_counts_file}{A file path (string) to a file detailing the number of reads in each bin
This file must have columns: "chromosome", "start position", "end position"
and "number of reads in region". i.e. a BED3+1 file where the final column
is the number of reads in the region}

\item{beta_threshold}{A numeric value that determines the significance threshold required for
a bin to be densely populated with reads. Defaults to 0.001}
}
\value{
A data.table detailing which bins in the input file are densely
populated with methylated sites. The data.table has columns:
"chromosome", "start position", "end position", "number of reads in region"
and "is densely methylated" (Boolean value)
}
\description{
Using the beta distribution, this finds bins with a higher
than expected density of reads.
}
\examples{
determine_dense_bins("path/to/bin_counts.bed", 200, 0.0001)

       chr start   end count density densely_methylated
    <char> <int> <int> <int>   <num>              <num>
 1:   chr1     0   200     3   0.015                  0
 2:   chr1   200   400     4   0.020                  0
 3:   chr1   400   600     0   0.000                  0
 4:   chr1   600   800     1   0.005                  0
 5:   chr1   800  1000   199   0.995                  1
 ...

}
