% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_rate.R
\name{estimate_error_rate}
\alias{estimate_error_rate}
\title{Estimate the Error Rate of Methylation Data}
\usage{
estimate_error_rate(reference_set_path)
}
\arguments{
\item{reference_set_path}{A file path (string) to a file detailing
CpG positions in the epigenome that do not exhibit methylation.
The file is expected to have columns: "read depth at base pair position"
and "percentage of such reads called as methylated", e.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ 34    5.9
 30    5
 42    4.4
 51    5.9
 22    0
}\if{html}{\out{</div>}}}
}
\value{
An estimated error rate (numeric) expressed as a proportion (0 to 1).
}
\description{
This function estimates the error rate of methylation data
using predicted unmethylated base pair positions. Any methylation signal
in the input file is considered incorrect and contributes to the estimated
error rate.
}
\details{
The input file should only contain positions with a low percentage
of reads called as methylated. By default, ChromBinarize uses a reference
set where the read depth is at least \strong{500} and the percentage of reads
being methylated is less than \strong{5\%}. It can also use only CpGs in CGIs,
which are typically unmethylated regions of the genome.

\strong{Warning}: If you are too stringent on unmethylated positions, the
estimated error rate may approach \strong{0}.
This may not be helpful. Aim for a larger error rate that is also accurate.
}
\examples{
# Estimate error rate using a TSV file
estimate_error_rate("path/to/reference_set.tsv")

# Estimate error rate using a CSV file
estimate_error_rate("path/to/reference_set.csv")

}
