% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_densely_methylated_bins.R
\name{get_beta_parameters}
\alias{get_beta_parameters}
\title{Estimates Shape Parameters of Beta Distribution}
\usage{
get_beta_parameters(bin_densities)
}
\arguments{
\item{bin_densities}{A numeric vector, detailing the collection of read densities across all
genomic regions considered.}
}
\value{
A numeric vector containing estimated parameters shape1 and shape2
}
\description{
Using the fitdistrplus package, the parameters for the beta
distribution (lovingly called shape1 and shape2) are estimated using
the reported density of reads in each window.
}
\examples{
 bin_densities <- c(0.001, 0.09, 0.1, 0.001, 0.002)
 get_beta_parameters(bin_densities)
 [1] 0.3834632 9.7126634
}
