% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_densely_methylated_bins.R
\name{is_densely_methylated}
\alias{is_densely_methylated}
\title{Determine if a Bin is Densely Methylated}
\usage{
is_densely_methylated(bin_density, shape1, shape2, threshold = 0.001)
}
\arguments{
\item{bin_density}{The density of the bin being inspected (numeric)}

\item{shape1, }{shape2
Beta distriubution parameters (numeric)}

\item{threshold}{The value required for the p-value from the beta distribution to be
considered significant}
}
\value{
A Boolean value (represented by a 0 or 1) where 1 indicates that
the bin density provided is significantly above what is expected (and 0
is the opposite of course)
}
\description{
Using the beta distribution and the given bin density the upper
tail of the beta distibution is calculated and compared against the given
threshold.
}
\examples{
 # A low density likely returns 0
 is_densely_methylated(0.001, 0.3, 9.5, 0.001)
 0

 # A high density likely returns 1
 is_densely_methylated(0.4, 0.3, 9.5, 0.001)
 1
}
