% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chromosome_sizes.R
\name{process_chromosome_sizes}
\alias{process_chromosome_sizes}
\title{Read and process a chromosome lengths file}
\usage{
process_chromosome_sizes(chromosome_sizes_file)
}
\arguments{
\item{chromosome_sizes_file}{A file path (string) to a text file detailing the length of each chromosome.
The file should contain two columns: chromosome name and chromosome length.}
}
\value{
A data.table with columns "chromsome" and "size" for each
chromosome given in the input file
}
\description{
Convert a file detailing the sizes of each chromosome
into a datatable with the correct column headings
}
\examples{
process_chromosome_sizes("path/to/chromosome_sizes.txt")

}
