% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedmethyl.R
\name{read_bedmethyl}
\alias{read_bedmethyl}
\title{Read in a Bedmethyl File}
\usage{
read_bedmethyl(bed_file_location)
}
\arguments{
\item{bed_file_location}{A file path (string) to the bedmethyl file to be
processed.}
}
\value{
A data.table with columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}
}
\description{
Converts a bedmethyl file into a data table that can be used
by many other functions in the pipeline.
}
\details{
This function also checks that your BEDmethyl file is of the
correct form. If your file has the incorrect number of columns or incorrect
column structure, a helpful error message will be returned instead.
}
\examples{
read_bedmethyl("path/to/bedmethyl_file.bed")

       chr start   end   name read_depth strand percent_methylation
    <char> <int> <int> <char>      <int> <char>               <num>
 1:   chr1   100   200      m         55      +                  20
 2:   chr1   300   400      h         20      +                  25
 3:   chr2   150   250      m         10      -                  20
 4:   chr2   500   600      h          1      +                   0
 5:   chr3   800   900      m         43      -                  49
 6:   chr3   950  1050      h         72      +                 100

}
