% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_densely_methylated_bins.R
\name{read_bin_counts_file}
\alias{read_bin_counts_file}
\title{Converts Bin Counts to Data Table}
\usage{
read_bin_counts_file(bin_counts_file)
}
\arguments{
\item{bin_counts_file}{A file path (string) to a file detailing the number of reads in each bin
This file must have columns: "chromosome", "start position", "end position"
and "number of reads in region". i.e. a BED3+1 file where the final column
is the number of reads in the region}
}
\value{
A data.table with the columns:
\itemize{
\item chr: Chromosome number (string, ex: "chr22")
\item start: Starting base pair position (integer)
\item end: Ending base pair position (integer)
\item count: Number of reads in the bin (integer)
}
}
\description{
Reads in a BED file detailing the number of reads in each
region as a data.table with the necessary column names for
\code{determine_dense_bins()}
}
\examples{
 read_bin_counts_file("path/to/bin_counts.bed")

       chr start   end count
    <char> <int> <int> <int>
 1:   chr1     0   200     3
 2:   chr1   200   400     4
 3:   chr1   400   600     0
 4:   chr1   600   800     1
 5:   chr1   800  1000   199
 ...
}
