% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_methylation_data.R
\name{read_methylation_data}
\alias{read_methylation_data}
\title{Read in a BEDMethyl File}
\usage{
read_methylation_data(bed_file_location)
}
\arguments{
\item{bed_file_location}{A file path (string) to the BEDMethyl file to be
processed.}
}
\value{
A data.table with columns: "chr" (chromosome name, string), "start",
(starting base pair position, int), "end" (ending base pair position, int),
"name" (m for 5mC and h for 5hmC), "read_depth" (read depth, int), "strand"
(Either "+", "-" or "." for unknown strand) and "percent_methylation"
(The percentage of such reads that were observed to be methylated, numeric)
}
\description{
Converts a BEDMethyl file into a data table that can be used
by many other functions in the pipeline.
}
\examples{
read_methylation_data("path/to/bedmethyl_file.bed")

       chr start   end   name read_depth strand percent_methylation
    <char> <int> <int> <char>      <int> <char>               <num>
 1:   chr1   100   200      m         55      +                  20
 2:   chr1   300   400      h         20      +                  25
 3:   chr2   150   250      m         10      -                  20
 4:   chr2   500   600      h          1      +                   0
 5:   chr3   800   900      m         43      -                  49
 6:   chr3   950  1050      h         72      +                 100

}
