% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_rate.R
\name{read_reference_set}
\alias{read_reference_set}
\title{Read in Reference Set}
\usage{
read_reference_set(reference_set_path)
}
\arguments{
\item{reference_set_path}{A file path (string) to a file detailing
CpG positions in the epigenome that do not exhibit methylation.
The file is expected to have columns: "read depth at base pair position"
and "percentage of such reads called as methylated", e.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ 34    5.9
 30    5
 42    4.4
 51    5.9
 22    0
}\if{html}{\out{</div>}}}
}
\value{
A data.table with columns "reads" (integer) and "percent_methylated"
(numerical).
}
\description{
Reads in your reference set into a data table with several
checks to ensure the file is of the correct form
}
\examples{
estimate_error_rate("path/to/reference_set.tsv")
}
