% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_5mc_5hmc.R
\name{combine_5mc_5hmc}
\alias{combine_5mc_5hmc}
\title{Combine the Signal from 5hmC and 5mC}
\usage{
combine_5mc_5hmc(comparison_bedmethyl)
}
\arguments{
\item{comparison_bedmethyl}{A data.table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth (integer)
\item percent_methylation: percentage of reads observed to be methylated
(numeric)
}}
}
\value{
A comparitive bedmethyl file with merged 5mC and 5hmC signal
}
\description{
Merges methylation signal for both "h" and "m". This is useful
when comparing ONT data to WGBS data, as WGBS doesn't discern between 5mC
and 5hmC (unlike modifided base called ONT data).
}
\examples{
# Read in a comparative bedmethyl file
read_comparison_bedmethyl("path/to/bedmethyl.bed")

# Combine signal
combine_5mc_5hmc(bedmethyl)

}
