% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_bed_files.R
\name{create_bins}
\alias{create_bins}
\title{Create genomic windows for a chromosome}
\usage{
create_bins(chromosome_name, chromosome_length, bin_size)
}
\arguments{
\item{chromosome_length}{The length of the given chromosome (integer)}

\item{bin_size}{The desired size of each region (integer)}

\item{chromsome}{The name of the chromsome (string). This must be of the
form "chrxyz" as ChromHMM expects this form.}
}
\value{
A data.table with columns "chromosome",
"starting base pair position" and "end base pair position"
}
\description{
Given a chromosome and its size this creates a data.table
that details regions of a fixed length spanning the chromosome
}
\examples{
 create_bins("chr22", "1000000", 200)

 Generates a data.table of the form:
 chr22	0	200
 chr22	200	400
 chr22	400	600
 ...
 chr22	999800	1000000
}
