% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_rate.R
\name{estimate_error_rate}
\alias{estimate_error_rate}
\title{Estimate the Error Rate of Methylation Data}
\usage{
estimate_error_rate(
  methylation_data,
  read_depth_threshold = 30,
  percent_threshold = 5
)
}
\arguments{
\item{read_depth_threshold}{The minimum read depth to consider (integer).
Higher values are recommended. Defaults to 30.}

\item{percent_threshold}{The maximum percent methylation to consider
(numeric). Lower values are recommended. Defaults to 5.}
}
\value{
An estimated error rate (numeric) expressed as a probability
(0 to 1).
}
\description{
This function estimates the error rate of methylation data
using predicted unmethylated base pair positions. Any methylation signal
in the input file is considered incorrect and contributes to the estimated
error rate.
}
\details{
The input file should only contain positions with a low percentage
of reads called as methylated. By default, ChromBinarize uses a reference
set where the read depth is at least \strong{500} and the percentage of reads
being methylated is less than \strong{5\%}. It can also use only CpGs in CGIs,
which are typically unmethylated regions of the genome.

\strong{Warning}: If you are too stringent on unmethylated positions, the
estimated error rate will approach 0.
This may not be helpful. Aim for a larger error rate that is also accurate.
}
\examples{
estimate_error_rate(bedmethyl_data, 40, 5)
estimate_error_rate(bedmethyl_data, 500, 2)
estimate_error_rate(bedmethyl_data)

# Bad usage
estimate_error_rate(bedmethyl_data, 1, 5)
estimate_error_rate(bedmethyl_data, 30, 0.1)

}
