% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_bed_files.R
\name{process_chromosome_sizes}
\alias{process_chromosome_sizes}
\title{Read and process a chromosome lengths file}
\usage{
process_chromosome_sizes(chromosome_sizes_file)
}
\arguments{
\item{chromosome_sizes_file}{A file path (string) to a text file detailing the length of each chromosome.
The file should contain two columns: chromosome name and chromosome length.}
}
\value{
An integer vector with values equal to the given chromosome lengths.
Each entry has a name equal to the chromosome name given in the input file
}
\description{
Convert a file detailing the sizes of each chromosome
into a datatable with the correct column headings
}
\details{
It is assumed that the input file is coming from ChromHMM's
included files. However, if this is not the case, you will need to ensure
that chromosomes are given in the form "chr22", "chrX" etc. (chr prefixed)
}
\examples{
process_chromosome_sizes("path/to/chromosome_sizes.txt")
}
