% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedmethyl.R
\name{read_comparison_bedmethyl}
\alias{read_comparison_bedmethyl}
\title{Read in a comparison BEDMethyl File}
\usage{
read_comparison_bedmethyl(bed_file_location)
}
\arguments{
\item{bed_file_location}{A file path (string) to the comparison BEDMethyl
file to be processed.}
}
\value{
A data.table with columns: "chr" (chromosome name, string), "start",
(starting base pair position, int), "end" (ending base pair position, int),
"mark_name" (m for 5mC and h for 5hmC), "read_depth" (read depth, int),
"strand" (Either "+", "-" or "." for unknown strand) and
"percent_methylation" (The percentage of such reads that were observed to
be methylated, numeric)
}
\description{
Converts a comparison BEDMethyl file into a data table that
can be used by the comparison scripts in the pipeline.
}
\details{
This function also checks that your BEDmethyl file is of the
correct form. If your file has the incorrect number of columns or incorrect
column structure, a helpful error message will be returned instead.
}
\examples{
read_comparison_bedmethyl("path/to/bedmethyl_file.bed")

       chr start   end mark_name ONT_read_depth ONT_percent_methylation
    <char> <int> <int>    <char>          <int>                   <num>
 1:   chr1   100   200         m             12                     8.0
 2:   chr1   300   400         h             22                    13.6
 3:   chr2   150   250         m             32                    50.0
 4:   chr2   500   600         h             25                     0.0
 5:   chr3   800   900         m             11                     9.0
 6:   chr3   950  1050         h             12                     8.0
    BS_read_depth BS_percent_methylation
            <int>                  <num>
 1:            14                   20.0
 2:             3                   66.7
 3:            40                  100.0
 4:            35                  100.0
 5:            32                   50.0
 6:            12                  100.0

}
