% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedmethyl.R
\name{regions_have_positive_length}
\alias{regions_have_positive_length}
\title{Ensure bedmethyl Regions are Positive}
\usage{
regions_have_positive_length(methylation_data)
}
\arguments{
\item{methylation_data}{A data table with the following columns:
\itemize{
\item chr: chromosome name (string)
\item start: starting base pair position (integer)
\item end: ending base pair position (integer)
\item mark_name: "m" for 5mC and "h" for 5hmC (string)
\item read_depth: read depth for the site (integer)
\item strand: "+", "-" or "." (character)
\item percent_methylation: percentage of reads reported as methylated (numeric)
}}
}
\value{
A logical value. TRUE implies that all regions make sense. FALSE
indicates at least one region is nonsensicle
}
\description{
BED files should have start and end columns such that each
region is strictly positive. A region spanning bp:200 (start) to bp:100
(end) is nonsensicle.
}
\examples{
 # Good bedmethyl data (end > start)
 regions_make_sense(bedmethyl_data)
   TRUE

 # bedmethyl data, row 3 contains a start value of 200, and end value
 # of 100
 bedmethyl
       chr start   end   name read_depth strand percent_methylation
    <char> <int> <int> <char>      <int> <char>               <num>
 1:   chr1   200   100      m         55      +                  20

 regions_make_sense(bedmethyl_data)
   FALSE
}
